# expression: implement vectorized evaluation for `SubXXXAndYYY`