<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
Use go template to generate vectorized  11 SUBTIME functions from #12176 

- builtinSubDatetimeAndDurationSig

- builtinSubDatetimeAndStringSig

- builtinSubDurationAndDurationSig

- builtinSubDurationAndStringSig

- builtinSubStringAndDurationSig

- builtinSubStringAndStringSig

- builtinSubDateAndDurationSig

- builtinSubDateAndStringSig

- builtinSubTimeDateTimeNullSig

- builtinSubTimeStringNullSig

- builtinSubTimeDurationNullSig

### What is changed and how it works?





    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDatetimeAndDurationSig-VecBuiltinFunc-4               3999            293371 ns/op               0 B/op          0 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDatetimeAndDurationSig-NonVecBuiltinFunc-4            3459            339815 ns/op               0 B/op          0 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDatetimeAndStringSig-VecBuiltinFunc-4                  586           2057947 ns/op           97826 B/op       6012 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDatetimeAndStringSig-NonVecBuiltinFunc-4               568           2100321 ns/op           98006 B/op       6012 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDurationAndDurationSig-VecBuiltinFunc-4             114855             10493 ns/op               0 B/op          0 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDurationAndDurationSig-NonVecBuiltinFunc-4           32994             36492 ns/op               0 B/op          0 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDurationAndStringSig-VecBuiltinFunc-4                 1057           1147146 ns/op           97517 B/op       6093 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDurationAndStringSig-NonVecBuiltinFunc-4               694           1796017 ns/op           99094 B/op       6093 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubStringAndDurationSig-VecBuiltinFunc-4                  518           2241379 ns/op          188826 B/op       9155 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubStringAndDurationSig-NonVecBuiltinFunc-4               537           2221352 ns/op          188730 B/op       9155 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubStringAndStringSig-VecBuiltinFunc-4                    326           3704692 ns/op          278973 B/op      14604 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubStringAndStringSig-NonVecBuiltinFunc-4                 363           3348957 ns/op          278401 B/op      14604 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDateAndDurationSig-VecBuiltinFunc-4                   2476            431365 ns/op           95126 B/op       3919 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDateAndDurationSig-NonVecBuiltinFunc-4                2412            463982 ns/op           95126 B/op       3919 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDateAndStringSig-VecBuiltinFunc-4                      692           1679502 ns/op          182640 B/op       9654 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubDateAndStringSig-NonVecBuiltinFunc-4                   525           2236717 ns/op          185652 B/op       9655 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubTimeDateTimeNullSig-VecBuiltinFunc-4              13765372                83.4 ns/op             0 B/op          0 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubTimeDateTimeNullSig-NonVecBuiltinFunc-4              75445             15781 ns/op               0 B/op          0 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubTimeStringNullSig-VecBuiltinFunc-4                  397507              2921 ns/op               0 B/op          0 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubTimeStringNullSig-NonVecBuiltinFunc-4               177312              6657 ns/op               0 B/op          0 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubTimeDurationNullSig-VecBuiltinFunc-4              13544961                85.4 ns/op             0 B/op          0 allocs/op
    BenchmarkVectorizedBuiltinTimeFuncGenerated/builtinSubTimeDurationNullSig-NonVecBuiltinFunc-4             115204             10282 ns/op               0 B/op          0 allocs/op






### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test